// rest/client/src/utils/types.ts
export interface Passwords {
    newPassword: string;
    confirmPassword: string;
    oldPassword: string;
}
// ...
export interface FieldProps<V = any> {
    field: FieldInputProps<V>;
    form: FormikProps<V>;
    meta: FieldMetaProps<V>;
}